Fix sign compare warnings
sizeof() (gcc-11.3.0 on Ubuntu) evaluates to a long unsigned int but is
being compared agains int variables.  Cast sizeof() in these cases
to int to suppress the warning.

+++ src/code.c
@@ -117,7 +117,7 @@ fprintf(FF,"INODES2CODE start\n");
 
 		/* advance offset by size of code type */
 		codetype = in->code.type;
-		if ( codetype <= 0 || codetype >= sizeof(Codesize) ) {
+		if ( codetype <= 0 || codetype >= (int)sizeof(Codesize) ) {
 			eprint("Unknown IC_* (%d) in 2inst!?",codetype);
 		}
 		else {
